+++ morphoLogicEngine/src/morphologic_server/archetypes/base.py
@@ -1,10 +1,8 @@
 """API for saving and querrying data from DB"""
 
-import json
-
 from abc import ABC  # , abstractmethod
 from enum import Enum
-from typing import Optional, Type, Tuple, Union
+from typing import Optional, Type
 
 from sqlalchemy import select
 
@@ -16,7 +14,6 @@ from geoalchemy2.shape import from_shape  # , to_shape
 from shapely.geometry import Point, Polygon
 
 from morphologic_server import logger
-from morphologic_server.db.engine import DBAsyncSession
 from morphologic_server.db.models import (
     BaseDB,
     AccountDB,
@@ -29,10 +26,6 @@ from morphologic_server.db.models import (
     CharacterDB,
     STANDARD_LENGTH as _DB_STANDARD_LENGTH,
 )
-from morphologic_server.utils import search as search_utils
-
-DEFAULT_SPAWN_LOCATION = from_shape(Point(0, 0, 0), srid=3857)
-
 
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ EXCEPTIONS ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #
 class PermissionDeniedError(Exception):
@@ -52,368 +45,6 @@ def convert_object_location_to_point(ob):
     """Convert a location of an object to a Shapely Point."""
     ob.location = shape.to_shape(ob.location)
     return ob
-
-
-# ------------------------------------------------------------------------------------------------ #
-#                        8888888b.  888888b.
-#                        888  "Y88b 888  "88b
-#                        888    888 888  .88P
-#                        888    888 8888888K.
-#                        888    888 888  "Y88b
-#                        888    888 888    888
-#                        888  .d88P 888   d88P
-#                        8888888P"  8888888P"
-#
-#
-# 8888888888                         888    d8b
-# 888                                888    Y8P
-# 888                                888
-# 8888888 888  888 88888b.   .d8888b 888888 888  .d88b.  88888b.  .d8888b
-# 888     888  888 888 "88b d88P"    888    888 d88""88b 888 "88b 88K
-# 888     888  888 888  888 888      888    888 888  888 888  888 "Y8888b.
-# 888     Y88b 888 888  888 Y88b.    Y88b.  888 Y88..88P 888  888      X88
-# 888      "Y88888 888  888  "Y8888P  "Y888 888  "Y88P"  888  888  88888P'
-# ------------------------------------------------------------------------------------------------ #
-
-
-# ************************************************************************************************ #
-#                                             SEACHING                                             #
-# ************************************************************************************************ #
-async def find_account(account_name: str):
-    """
-    Finds account by name.
-    If "#<int>" is being searched, it will be searched by ID.
-
-    Args:
-        account_name (str): Name or #ID of the account to search for.
-
-    Returns:
-        Account: if found, None otherwise.
-    """
-    async with DBAsyncSession() as session:
-
-        if account_name.startswith("#"):
-            try:
-                account_id = int(account_name[1:])
-            except ValueError:
-                logger.warning(
-                    "Invalid account ID format: %s. Expected format: #[int].",
-                    account_name,
-                )
-                return None
-            stmt = select(AccountDB).where(AccountDB.id == account_id)
-
-        else:
-            # If account name is not a number, search by name
-            stmt = select(AccountDB).where(AccountDB.name == account_name)
-        result = await session.execute(stmt)
-        account = result.scalars().first()
-
-        if account:
-            return Account(account)
-        return None
-
-
-async def search(name_or_id: str, archetype: Type["Archetypes"]):
-    """Search for an object  by name or ID.
-    If "#[int]" is being searched, it will be searched by ID.
-
-    Args:
-        name_or_id (str): Name or I#ID of the object to search for.
-        archetype (Archetypes): The archetype class to search in.
-
-    Returns:
-        Object: if found, None otherwise.
-    """
-    model = archetype.linked_db_obj
-
-    async with DBAsyncSession() as session:
-        if name_or_id.startswith("#"):
-            try:
-                object_id = int(name_or_id[1:])
-            except ValueError:
-                logger.warning(
-                    "Invalid object ID format: %s. Expected format: #[int].",
-                    name_or_id,
-                )
-                return None
-            stmt = select(model).where(model.id == object_id)
-        else:
-            # If name_or_id is not a number, search by name
-            stmt = select(model).where(model.name == name_or_id)
-        result = await session.execute(stmt)
-        obj = result.scalars().first()
-
-        return archetype(obj) if obj else None
-
-
-async def simple_query(value, attribute: str, model: Type["Archetypes"]):
-    """Simple query to find rows from specified table by attribute.
-
-    Args:
-        value (any): the value to search for
-        attribute (str): an atttribute to search for
-        model (Archetypes): the model to search in
-
-    Returns:
-        Archetypes: Returns object or objects if found, None otherwise.
-    """
-    async with DBAsyncSession() as session:
-        stmt = select(model.linked_db_obj).where(
-            getattr(model.linked_db_obj, attribute) == value
-        )
-        result = await session.execute(stmt)
-        obj = result.scalars().all()
-
-        if obj:
-            return [model(o) for o in obj] if len(obj) > 1 else model(obj[0])
-        return None
-
-
-async def search_by_xy(x: float, y: float, archetype: Type["Archetypes"]):
-    """Search for an object by XY coordinates.
-
-    Args:
-        x (float): x coordinate
-        y (float): y coordinate
-        archetype (Archetypes child): The archetype class to search in.
-    Returns:
-        Object: if found, None otherwise.
-    """
-    point = from_shape(Point(x, y), srid=3857)
-    model = archetype.linked_db_obj
-    async with DBAsyncSession() as session:
-        stmt = select(model).where(
-            ST_DWithin(
-                ST_Force2D(model.location),
-                ST_Force2D(point),
-                0.1,
-            )
-        )
-        result = await session.execute(stmt)
-        obj = result.scalars().first()
-
-        return archetype(obj) if obj else None
-
-
-# def get_area_containing_point(session, x: float, y: float):
-#     point = Point(x, y)
-#     # point3d = Point(float(x), float(y), float(z))
-#     point_geom = shape.from_shape(point, srid=3857)
-#     stmt = select(AreaDB).where(ST_Intersects(AreaDB.polygon, point_geom))
-#     result = session.execute(stmt).scalars().all()
-#     return result
-
-
-# ************************************************************************************************ #
-#                                             CREATING                                             #
-# ************************************************************************************************ #
-async def create_account(name: str, email: str) -> Optional["Account"]:
-    """Creates an account in the database.
-
-    Returns:
-        Account: Archetypes Accheptor object if created, None otherwise.
-    """
-    async with DBAsyncSession() as session:
-        account = AccountDB(name=name, email=email, permission_level=2)
-        session.add(account)
-        await session.commit()
-        await session.refresh(account)
-        return Account(account)
-
-
-async def create_character_and_soul(
-    account_id,
-    name: str,
-    description: str = "",
-    location=None,
-    permission_level: int = 2,
-) -> Optional[list["Character", "CharacterSoul"]]:
-    """Creates a character and a character soul in the database.
-
-    Returns:
-        List[Character, CharacterSoul]: Archetypes Character and CharacterSoul objects if created,
-            None otherwise.
-    """
-    if location is None:
-        location = DEFAULT_SPAWN_LOCATION
-
-    async with DBAsyncSession() as session:
-        character = CharacterDB(
-            name=name,
-            description=description,
-            location=location,
-            # soul_id=character_soul.id,
-        )
-        character_soul = CharacterSoulDB(
-            aura=0,
-            account_id=account_id,
-            permission_level=permission_level,
-            bound_character=character,
-        )
-        session.add(character_soul)
-        await session.commit()
-        await session.refresh(character_soul)
-        await session.refresh(character)
-
-        return [Character(character), CharacterSoul(character_soul)]
-
-
-async def create_character(
-    name: str,
-    soul: "CharacterSoul",
-    description: str = "",
-    location=None,
-    attributes: dict = None,
-    container: "GameObject" = None,
-    stored: list["GameObject"] = None,
-) -> Optional["Character"]:
-    """Creates a character in the database.
-
-    Returns:
-        Character: Archetypes Character object if created, None otherwise.
-    """
-    if location is None:
-        location = DEFAULT_SPAWN_LOCATION
-
-    async with DBAsyncSession() as session:
-        character = CharacterDB(
-            name=name,
-            soul=soul,
-            description=description,
-            location=location,
-            attributes=attributes,
-            container=container,
-            stored=stored,
-        )
-        session.add(character)
-        await session.commit()
-        await session.refresh(character)
-        return Character(character)
-
-
-async def create_or_edit_terrain(
-    x: int, y: int, z: int = 0, terrain_type: TerrainType = TerrainType.SOIL
-) -> Type["Terrain"]:
-    """Add or edit terrain in DB
-    Args:
-        x (int): x coordinate
-        y (int): y coordinate
-        z (int, optional): z coordinate. Defaults to 0.
-        terrain_type (TerrainType, optional): type of terrain. Defaults to TerrainType.SOIL.
-
-    Returns:
-        Terrain: Created or updated terrain object.
-    """
-    async with DBAsyncSession() as session:
-        point3d = Point(float(x), float(y), float(z))
-        point_geom = shape.from_shape(point3d, srid=3857)
-
-        # First check if the terrain already exists
-        stmt = select(TerrainDB).where(
-            ST_DWithin(
-                ST_Force2D(TerrainDB.location),
-                ST_Force2D(point_geom),
-                0.1,
-            )
-        )
-        result = await session.execute(stmt)
-        terrain = result.scalars().first()
-
-        if terrain:
-            # Terrain already exists, update it
-            old_point = shape.to_shape(terrain.location)
-            new_point = Point(old_point.x, old_point.y, float(z))
-
-            terrain.location = shape.from_shape(new_point, srid=3857)
-            terrain.type = terrain_type
-        else:
-            # Terrain doesn't exist, create a new one
-
-            terrain = TerrainDB(location=point_geom, type=terrain_type)
-            session.add(terrain)
-
-        await session.commit()
-        await session.refresh(terrain)
-        return Terrain(terrain)
-
-
-async def create_area(
-    polygon: list[Tuple[float, float]], name: str, description: str, priority: int = 0
-) -> Optional["Area"]:
-    """Creates an area in the database.
-
-    Returns:
-        Area: Archetypes Area object if created, None otherwise.
-    """
-    async with DBAsyncSession() as session:
-        polygon = Polygon(polygon)
-        area = AreaDB(
-            polygon=shape.from_shape(polygon, srid=3857),
-            name=name,
-            description=description,
-            priority=priority,
-        )
-        session.add(area)
-        await session.commit()
-        await session.refresh(area)
-        return Area(area)
-
-
-async def create_game_object(
-    name: str,
-    description: str = "",
-    location=None,
-    attributes: dict = None,
-    container: "GameObject" = None,
-    stored: Union[list["GameObject"], "GameObject"] = None,
-) -> Optional["GameObject"]:
-    """Creates a game object in the database.
-
-    Returns:
-        GameObject: Archetypes GameObject object if created, None otherwise.
-    """
-    if location is None:
-        location = DEFAULT_SPAWN_LOCATION
-    else:
-        location = from_shape(location, srid=3857)
-
-    kwargs = {}
-    if container is not None:
-        container = container.db_obj
-        kwargs["container"] = container
-    if stored is not None:
-        if not isinstance(stored, list):
-            stored = [stored]
-        stored = [stored.db_obj for obj in stored]
-        kwargs["stored"] = stored
-
-    async with DBAsyncSession() as session:
-        game_object = GameObjectDB(
-            name=name,
-            description=description,
-            location=location,
-            attributes=attributes,
-            **kwargs,
-        )
-        session.add(game_object)
-        await session.commit()
-        await session.refresh(game_object)
-        return GameObject(game_object)
-
-
-async def authenticate(username: str, password: str) -> Optional["Character"]:
-    """Return the bound Character if username + password are valid, else None.
-
-    Lookup chain: Account.name → account.character_souls[0] → soul.bound_character.
-    """
-    account = await find_account(username)
-    if account is None or not account.check_password(password):
-        return None
-    souls = account.character_souls
-    if not souls:
-        return None
-    return souls[0].bound_character
 
 
 # ------------------------------------------------------------------------------------------------ #
@@ -437,6 +68,7 @@ class Archetypes(ABC):
 
     linked_db_obj = BaseDB
     _db_obj = None
+    _sessionmaker = None  # Set at startup by MorphoLogicHeart
 
     @property
     def db_obj(self):
@@ -488,7 +120,7 @@ class Archetypes(ABC):
     async def save(self):
         """Save the object to the database."""
         if self._db_obj is not None:
-            async with DBAsyncSession() as session:
+            async with Archetypes._sessionmaker() as session:
                 # Attributes of this class always mirror _db_obj attributes
                 # So we can just pass the _db_obj to the SQLAlchemy session
                 session.add(self._db_obj)
@@ -498,7 +130,7 @@ class Archetypes(ABC):
     # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ Refresh ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #
     async def refresh(self):
         """Refresh the object from the database."""
-        async with DBAsyncSession() as session:
+        async with Archetypes._sessionmaker() as session:
             refreshed = await session.execute(
                 select(self._db_obj.__class__).where(
                     self._db_obj.__class__.id == self._db_obj.id
@@ -511,65 +143,11 @@ class Archetypes(ABC):
     # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ Delete ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #
     async def delete(self):
         """Deletes the object from the database."""
-        async with DBAsyncSession() as session:
+        async with Archetypes._sessionmaker() as session:
             await session.delete(self._db_obj)
             await session.commit()
             # self._db_obj = None
         return {"success": True, "message": "Calling object deleted from database."}
-
-    # ******************************************************************************************** #
-    #                                           WRAPPERS                                           #
-    # ******************************************************************************************** #
-    async def find_account(self, account_name: str):
-        """
-        Finds account by name.
-        If "#<int>" is being searched, it will be searched by ID.
-
-        Args:
-            account_name (str): Name or #ID of the account to search for.
-
-        Returns:
-            Account: if found, None otherwise.
-        """
-        return await find_account(account_name)
-
-    async def search(self, name_or_id: str, archetype: Type["Archetypes"]):
-        """Search for an object  by name or ID.
-        If "#[int]" is being searched, it will be searched by ID.
-
-        Args:
-            name_or_id (str): Name or I#ID of the object to search for.
-            archetype (Archetypes): The archetype class to search in.
-
-        Returns:
-            Object: if found, None otherwise.
-        """
-        return await search(name_or_id, archetype)
-
-    async def simple_query(self, value, attribute: str, model: Type["Archetypes"]):
-        """Simple query to find rows from specified table by attribute.
-
-        Args:
-            value (any): the value to search for
-            attribute (str): an atttribute to search for
-            model (Archetypes): the model to search in
-
-        Returns:
-            Archetypes: Returns object or objects if found, None otherwise.
-        """
-        return await simple_query(value, attribute, model)
-
-    async def search_by_xy(self, x: float, y: float, archetype: Type["Archetypes"]):
-        """Search for an object by XY coordinates.
-
-        Args:
-            x (float): x coordinate
-            y (float): y coordinate
-            archetype (Archetypes child): The archetype class to search in.
-        Returns:
-            Object: if found, None otherwise.
-        """
-        return await search_by_xy(x, y, archetype)
 
 
 #        d8888                                            888
@@ -785,128 +363,6 @@ class CharacterSoul(Archetypes):
     #                                            METHODS                                           #
     # ******************************************************************************************** #
 
-    # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ <Create> Wrappers ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #
-    async def create_account(self, name: str, email: str) -> Optional["Account"]:
-        """Creates an account in the database.
-
-        Returns:
-            Account: Archetypes Accheptor object if created, None otherwise.
-        """
-        self.check_permissions()
-
-        return await create_account(name=name, email=email)
-
-    async def create_character_and_soul(
-        self,
-        account_id,
-        name: str,
-        description: str = "",
-        location=None,
-        permission_level: int = 2,
-    ) -> Optional[list["Character", "CharacterSoul"]]:
-        """Creates a character and a character soul in the database.
-
-        Returns:
-            List[Character, CharacterSoul]: Archetypes Character and CharacterSoul objects if created,
-                None otherwise.
-        """
-        self.check_permissions()
-
-        return await create_character_and_soul(
-            account_id=account_id,
-            name=name,
-            description=description,
-            location=location,
-            permission_level=permission_level,
-        )
-
-    async def create_character(
-        self,
-        name: str,
-        soul: "CharacterSoul",
-        description: str = "",
-        location=None,
-        attributes: dict = None,
-        container: "GameObject" = None,
-        stored: list["GameObject"] = None,
-    ) -> Optional["Character"]:
-        """Creates a character in the database.
-
-        Returns:
-            Character: Archetypes Character object if created, None otherwise.
-        """
-        self.check_permissions()
-
-        return await create_character(
-            name=name,
-            soul=soul,
-            description=description,
-            location=location,
-            attributes=attributes,
-            container=container,
-            stored=stored,
-        )
-
-    async def create_or_edit_terrain(
-        self, x: int, y: int, z: int = 0, terrain_type: TerrainType = TerrainType.SOIL
-    ) -> Type["Terrain"]:
-        """Add or edit terrain in DB
-        Args:
-            x (int): x coordinate
-            y (int): y coordinate
-            z (int, optional): z coordinate. Defaults to 0.
-            terrain_type (TerrainType, optional): type of terrain. Defaults to TerrainType.SOIL.
-
-        Returns:
-            Terrain: Created or updated terrain object.
-        """
-        self.check_permissions()
-
-        return await create_or_edit_terrain(x=x, y=y, z=z, terrain_type=terrain_type)
-
-    async def create_area(
-        self,
-        polygon: list[Tuple[float, float]],
-        name: str,
-        description: str,
-        priority: int = 0,
-    ) -> Optional["Area"]:
-        """Creates an area in the database.
-
-        Returns:
-            Area: Archetypes Area object if created, None otherwise.
-        """
-        self.check_permissions()
-
-        return await create_area(
-            polygon=polygon, name=name, description=description, priority=priority
-        )
-
-    async def create_game_object(
-        self,
-        name: str,
-        description: str = "",
-        location=None,
-        attributes: dict = None,
-        container: "GameObject" = None,
-        stored: Union[list["GameObject"], "GameObject"] = None,
-    ) -> Optional["GameObject"]:
-        """Creates a game object in the database.
-
-        Returns:
-            GameObject: Archetypes GameObject object if created, None otherwise.
-        """
-        self.check_permissions()
-
-        return await create_game_object(
-            name=name,
-            description=description,
-            location=location,
-            attributes=attributes,
-            container=container,
-            stored=stored,
-        )
-
     # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ DELETE ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #
     def permission_level_is(self, permission_levels: int | tuple[int]) -> bool:
         """Check if the soul has the required permission level.
@@ -953,7 +409,7 @@ class CharacterSoul(Archetypes):
         """
         # First we check if the permission level of the souls is 0 or 1 (admin or builder)
         self.check_permissions()
-        async with DBAsyncSession() as session:
+        async with Archetypes._sessionmaker() as session:
             await session.delete(target_object.db_obj)
             await session.commit()
         return {"success": True, "message": "Target object deleted."}
@@ -1029,7 +485,7 @@ class Terrain(Archetypes):
             return None
 
         point = from_shape(Point(x, y), srid=3857)
-        async with DBAsyncSession() as session:
+        async with Archetypes._sessionmaker() as session:
             stmt = select(TerrainDB).where(
                 ST_DWithin(TerrainDB.location, point, distance_m)
             )
@@ -1218,7 +674,7 @@ class GameObject(Archetypes):
         """Return container linked to this game object"""
         if not self.container_id:
             return None
-        async with DBAsyncSession() as session:
+        async with Archetypes._sessionmaker() as session:
             result = await session.execute(
                 select(GameObjectDB).where(GameObjectDB.id == self.container_id)
             )
@@ -1228,9 +684,7 @@ class GameObject(Archetypes):
     # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ Stored ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #
     async def stored(self):
         """Return game objects stored in this game object"""
-        # container = search(f"#{self._db_obj.container_id}", GameObject)
-        # return container if container else None
-        async with DBAsyncSession() as session:
+        async with Archetypes._sessionmaker() as session:
             result = await session.execute(
                 select(GameObjectDB).where(GameObjectDB.container_id == self.id)
             )
@@ -1299,10 +753,7 @@ class Character(GameObject):
     #                                            METHODS                                           #
     # ******************************************************************************************** #
     async def get_areas_im_in(self):
-        # point = Point(x, y)
-        # point3d = Point(float(x), float(y), float(z))
-        # point_geom = shape.from_shape(point, srid=3857)
-        async with DBAsyncSession() as session:
+        async with Archetypes._sessionmaker() as session:
             stmt = select(AreaDB).where(
                 ST_Intersects(AreaDB.polygon, ST_Force2D(self._db_obj.location))
             )
@@ -1321,10 +772,10 @@ class Character(GameObject):
         # sorted_list = sorted(my_list, key=lambda obj: obj.priority)
         return areas[0] if areas else None
 
-    async def get_surrounding_description(self):
+    async def get_surrounding_description(self, memory):
         """Get description of the area this character is in."""
         area = await self.get_area_im_in()
-        objects = await search_utils.get_objects_in_proximity(self)
+        objects = await memory.get_objects_in_proximity(self)
         game_objects = objects.get("game_objects", [])
         characters = objects.get("characters", [])
         characters_str = "\n".join([f"{char.name} ({char.id})" for char in characters])
@@ -1332,9 +783,3 @@ class Character(GameObject):
 
         test_text = f'Character {self.name} is in area "{area.name}" \n Characters around:\n {characters_str} \n\n Objects around:\n {game_objects_str}'
         return test_text
-
-# ------------------------------------------------------------------------------------------------ #
-# Tymczasowo, żeby móc w shell'u korzystać z self
-async def get_self():
-    """Get self object for testing purposes."""
-    return await search("MoonyTheDream", Character)
