+++ morphoLogicEngine/src/morphologic_server/scripts/seed.py
@@ -16,11 +16,7 @@ import asyncio
 from shapely.geometry import Point
 
 from morphologic_server.archetypes.base import (
-    search,
-    find_account,
-    create_area,
-    create_game_object,
-    create_character_and_soul,
+    Archetypes,
     Character,
     GameObject,
 )
@@ -35,13 +31,13 @@ def _pt(x: float, y: float, z: float = 0.0) -> Point:
     return Point(x, y, z)
 
 
-async def _ensure_object(name: str, description: str, location: Point, container=None):
+async def _ensure_object(memory, name: str, description: str, location: Point, container=None):
     """Create a game object only if one with this name doesn't already exist."""
-    existing = await search(name, GameObject)
+    existing = await memory.search(name, GameObject)
     if existing:
         logger.info("Seed: object '%s' already exists — skipping.", name)
         return existing
-    obj = await create_game_object(
+    obj = await memory.create_game_object(
         name=name,
         description=description,
         location=location,
@@ -55,15 +51,22 @@ async def _ensure_object(name: str, description: str, location: Point, container
 # ---------------------------------------------------------------------------
 
 async def seed():
+    from morphologic_server.db.engine import create_sessionmaker
+    from morphologic_server.db.memory import Memory
+    from morphologic_server.archetypes.base import Area
+    from morphologic_server.config import ServerSettings
+
+    settings = ServerSettings()
+    sessionmaker = create_sessionmaker(settings.DB_ADDRESS)
+    Archetypes._sessionmaker = sessionmaker
+    memory = Memory(sessionmaker)
+
     logger.info("Seeding database…")
 
     # ── Areas ────────────────────────────────────────────────────────────── #
-    from morphologic_server.archetypes.base import simple_query, Area
-    from morphologic_server.db.models import AreaDB
-
-    existing_areas = await simple_query("Mały, wierzbowy lasek.", "name", Area)
+    existing_areas = await memory.simple_query("Mały, wierzbowy lasek.", "name", Area)
     if not existing_areas:
-        await create_area(
+        await memory.create_area(
             polygon=[(-10, -10), (10, -10), (10, 10), (-10, 10), (-10, -10)],
             name="Mały, wierzbowy lasek.",
             description="Cichy lasek nad rzeką, gdzie rosną stare wierzby.",
@@ -73,9 +76,9 @@ async def seed():
     else:
         logger.info("Seed: area 'Mały, wierzbowy lasek.' already exists — skipping.")
 
-    existing_polana = await simple_query("Polana", "name", Area)
+    existing_polana = await memory.simple_query("Polana", "name", Area)
     if not existing_polana:
-        await create_area(
+        await memory.create_area(
             polygon=[(5, 5), (20, 5), (20, 20), (5, 20), (5, 5)],
             name="Polana",
             description="Otwarta polana z miękką trawą i dzikim kwieciem.",
@@ -87,35 +90,35 @@ async def seed():
 
     # ── Game objects spread around spawn ─────────────────────────────────── #
     # Top-level objects (no container)
-    await _ensure_object("Stary dąb",       "Rozłożysty dąb z wyraźną dziuplą.",      _pt( 3,  4))
-    await _ensure_object("Kamień graniczny", "Omszały kamień z wyrytym znakiem.",      _pt(-4,  2))
-    await _ensure_object("Wiadro",          "Drewniane wiadro bez ucha.",              _pt( 1, -5))
-    await _ensure_object("Ognisko",         "Dopalające się ognisko.",                 _pt(-2, -2))
-    await _ensure_object("Skrzynia",        "Ciężka drewniana skrzynia.",              _pt( 7,  1))
-    await _ensure_object("Pędzel Moony'ego","Cieniutki pędzel z czarną rączką.",       _pt(-1, -3))
+    await _ensure_object(memory, "Stary dąb",       "Rozłożysty dąb z wyraźną dziuplą.",      _pt( 3,  4))
+    await _ensure_object(memory, "Kamień graniczny", "Omszały kamień z wyrytym znakiem.",      _pt(-4,  2))
+    await _ensure_object(memory, "Wiadro",          "Drewniane wiadro bez ucha.",              _pt( 1, -5))
+    await _ensure_object(memory, "Ognisko",         "Dopalające się ognisko.",                 _pt(-2, -2))
+    await _ensure_object(memory, "Skrzynia",        "Ciężka drewniana skrzynia.",              _pt( 7,  1))
+    await _ensure_object(memory, "Pędzel Moony'ego","Cieniutki pędzel z czarną rączką.",       _pt(-1, -3))
 
     # Coins placed individually at different spots
     coin_pouch = await _ensure_object(
-        "Coin Pouch", "Skórzana sakiewka z monetami.", _pt(3, -1)
+        memory, "Coin Pouch", "Skórzana sakiewka z monetami.", _pt(3, -1)
     )
-    await _ensure_object("Copper Coin", "Miedziana moneta.",  _pt(1,  0))
-    await _ensure_object("Silver Coin", "Srebrna moneta.",    _pt(-2, 1))
-    await _ensure_object("Gold Coin",   "Złota moneta.",      _pt(4, -2))
+    await _ensure_object(memory, "Copper Coin", "Miedziana moneta.",  _pt(1,  0))
+    await _ensure_object(memory, "Silver Coin", "Srebrna moneta.",    _pt(-2, 1))
+    await _ensure_object(memory, "Gold Coin",   "Złota moneta.",      _pt(4, -2))
 
     # ── Extra NPC character ───────────────────────────────────────────────── #
     # Only create if an account named "Another" exists; skip otherwise.
-    another_account = await find_account("Another")
+    another_account = await memory.find_account("Another")
     if another_account:
-        existing_char = await search("AnotherCharacter", Character)
+        existing_char = await memory.search("AnotherCharacter", Character)
         if not existing_char:
-            await create_character_and_soul(
+            await memory.create_character_and_soul(
                 account_id=another_account.id,
                 name="AnotherCharacter",
                 description="A wandering soul.",
                 location=None,  # will use DEFAULT_SPAWN_LOCATION; move below
             )
             # Move them to (2, 3) so they're visible but distinct
-            char = await search("AnotherCharacter", Character)
+            char = await memory.search("AnotherCharacter", Character)
             if char:
                 char.location = (2.0, 3.0, 0.0)
                 await char.save()
